minor adjustments to Yibin's profile page
