+++ src/assets/main/_main.scss
@@ -4,7 +4,7 @@
 // color theme
 $colors: (
     primary: #12606A,
-    secondary: #12606A,
+    secondary: #B8CED0,
     black: #000000,
     white: #FFFFFF,
     light1: #DFE2EA,
