+++ src/components/Footer/Footer.js
@@ -1,8 +1,5 @@
 import React from "react";
 import styles from "./Footer.module.scss";
-import Link from "@material-ui/core/Link";
-import Typography from "@material-ui/core/Typography";
-import Button from "@material-ui/core/Button";
 
 const Footer = () => {
   return <footer className={styles.footer}></footer>;
