+++ src/pages/HomePage/HomePage.js
@@ -1,7 +1,6 @@
 import React from "react";
 import styles from "./HomePage.module.scss";
 import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
 import Footer from "../../components/Footer/Footer";
 import Header from "../../components/Header/Header";
 import TextField from "@material-ui/core/TextField";
