--- src/pages/ProfilePage/ProfileSubPage/AboutUs.js
@@ -1,19 +1,0 @@
-import React, { Component } from "react";
-
-const AboutUs = (props) => {
-  return (
-    <div>
-      <h2>About Us</h2>
-      <p>Here is information about our company/student organization!</p>
-      <ol>
-        <li>Our Mission</li>
-        <li>Goals</li>
-        <li>Events running this year</li>
-        <li>How to help</li>
-        <li>What we do</li>
-      </ol>
-    </div>
-  );
-};
-
-export default AboutUs;
